+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Changelog
 
 * View load instrumentation
   [65](https://github.com/bugsnag/bugsnag-flutter-performance/pull/65)
+  
+* Auto-inject traceparent headers into HTTP requests [73](https://github.com/bugsnag/bugsnag-flutter-performance/pull/73)
 
 ## 1.0.0 (2024-04-11)